requirements.txt
pyarrow
pandas
numpy
streamlit
plotly
pillow